==> scripts/04_MCPcounter.Rmd <==
---
title: "Untitled"
author: "yanan zhao"
date: '2022-06-13'
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

```

## R Markdown

This is an R Markdown document. Markdown is a simple formatting syntax for authoring HTML, PDF, and MS Word documents. For more details on using R Markdown see <http://rmarkdown.rstudio.com>.

When you click the **Knit** button a document will be generated that includes both content as well as the output of any embedded R code chunks within the document. You can embed an R code chunk like this:

Note that the `echo = FALSE` parameter was added to the code chunk to prevent printing of the R code that generated the plot.



```{r, include=FALSE}
install.packages(c("devtools","curl")) ##Installs devtools and the MCPcounter dependancy 'curl'
library(devtools)
install_github("ebecht/MCPcounter",ref="master", subdir="Source")
library(tidyverse)
library(MCPcounter)
library("ggrepel")
```



```{r}
expressionfinal = read.table("../data/gene_expression.tsv",header = TRUE,row.names = 1)

pheno = read.table('../data/pheno.txt',header = TRUE)
```

```{r}
#matchlist= read.table('../data/probeID_gene.tsv', header = TRUE)

#matchlist = matchlist %>% rename(names = probeID)
```

```{r}
#p2g <- split(matchlist$ID, matchlist$gene_symbol)
#gene_exp <- t(sapply(p2g, function(x) colSums(expr[rownames(expr) %in% x,])))


#probe_exp[which.max(rowSums(probe_exp[rownames(probe_exp) %in% x,])),]
```




```{r}
#mcp = left_join(expression, matchlist, by = "names") 
```


```{r}

#expressionall = mcp %>% select(-51) %>% filter(!grepl('---',GeneSymbol)) %>% na.omit() 


#expressionfinal = aggregate(expressionall[, 1:50], list(expressionall$GeneSymbol), mean)


#expressionfinal = expressionfinal %>% remove_rownames %>% column_to_rownames(var="Group.1") 

```





```{r}
# immune cells expression level

expressionfinal = read.table("../data/gene_expression.tsv",header = TRUE) %>% remove_rownames %>% column_to_rownames(var="gene_symbol") 



RESULT = MCPcounter.estimate(expressionfinal,featuresType='HUGO_symbols')


heatmap(as.matrix(RESULT),col=colorRampPalette(c("blue","white","red"))(100) ) 
```


```{r}

P = test_for_infiltration(RESULT,platform=c("HG1"))
P
#expressionexclude = expressionfinal %>% filter(rownames(expressionfinal) #%in% matchlist[["GeneSymbol"]])

```



```{r}
pheno = pheno %>% mutate(age = str_remove_all(age, "y")) %>% unite(type, tissue:timepoint) %>% mutate(type = gsub("_pretreatment", "", type)) %>% mutate(type = gsub("eeks after start of", "", type))

pheno$age = sapply(pheno$age, as.numeric)
```


```{r}
pca <- prcomp(t(RESULT))

## Check out how much variance is explained
summary(pca)

## Plot PCA
df <- data.frame(PC1 = pca$x[,1], PC2 = pca$x[,2])

df$condition <- pheno$type
ggplot(df, aes(x = PC1, y = PC2, label = rownames(df), color = condition)) +
    geom_text(size = 3)


```




```{r}
MCPtem = data.frame(t(RESULT))
MCPtem$res = pheno$treatment_response
    
ggplot(MCPtem, aes(x = MCPtem$res, y =MCPtem$Endothelial.cells)) +
  geom_boxplot()

```


```{r}
info = pheno 
info$sample = rownames(df)
```


```{r}
BiocManager::install("ConsensusClusterPlus")
library(ConsensusClusterPlus)
```



```{r}
library(corrplot)

a = data.frame(t(RESULT))

corrplot(cor(a), type = "upper", order = "hclust", 
         tl.col = "black", tl.srt = 45)
```



```{r}
cluster = ConsensusClusterPlus(RESULT,maxK=6,reps=50,pItem=0.8,pFeature=1, title= 'mcp_cluster',clusterAlg="hc",distance="pearson",seed=1262118388.71279,plot="png")


calcICL(cluster,title="untitled_consensus_cluster",plot=NULL,writeTable=FALSE)

```

```{r}
x = data.frame(cluster[[3]]$consensusClass) 

x$treat=pheno$treatment_response 
x$type=pheno$type
x

```

```{r}
pca <- prcomp(t(RESULT))

## Check out how much variance is explained
summary(pca)

## Plot PCA
df <- data.frame(PC1 = pca$x[,1], PC2 = pca$x[,2])

df$condition <- as.character(x$cluster..3...consensusClass)
ggplot(df, aes(x = PC1, y = PC2, label = rownames(df), color = condition)) +
    geom_text(size = 3)

```




```{r}
write.table(RESULT, "../data/MCPcounter.txt", sep = "\t", row.names = TRUE, quote = FALSE)

#write.table(expressionfinal, "../data/GeneSymbol_expr.txt", sep = "\t", row.names = TRUE, quote = FALSE)
```
